refactor: :truck: Coherent file structure
